pyproject.toml
[tool.poetry]
name = "livecryptopricedjango"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
python-dotenv = "^1.0.1"
celery = "^5.4.0"
redis = "^5.0.7"
certifi = "^2024.7.4"
requests = "^2.32.3"
djangorestframework = "^3.15.2"
dj-database-url = "^2.2.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^22.0.0"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
